docs(viewer-docs): add AnnotationViewer, Drawer docs
